Fix CLI imports and tests
