+++ mypy.ini
@@ -10,6 +10,9 @@ ignore_errors = True
 [mypy-ume_cli]
 ignore_errors = True
 
+[mypy-ume.cli.*]
+ignore_errors = True
+
 [mypy-ume_client.*]
 ignore_errors = True
